feat: support num_tokens 0 and shape 0 input tensor
* support cpp case nullptr inpTokBuf
* fix -O0 maxTokenSize overflow bug
* add HIP_RUNTIME_CHECK_WITH_BACKTRACE

+++ src/ops/dispatch_combine/dispatch_combine.cpp
@@ -72,7 +72,8 @@ mori::application::SymmMemObjPtr ShmemMallocAndReturnMemObjPtr(size_t size, unsi
 }
 
 void EpDispatchCombineHandle::InitializeShmemBuf() {
-  size_t maxTokenSize = config.MaxNumTokensToRecv() * config.hiddenDim * config.maxTokenTypeSize;
+  size_t maxTokenSize = static_cast<ssize_t>(config.MaxNumTokensToRecv()) * config.hiddenDim *
+                        config.maxTokenTypeSize;
   shmemInpTokMemObj = ShmemMallocAndReturnMemObjPtr(maxTokenSize, hipDeviceMallocUncached);
   shmemOutTokMemObj = ShmemMallocAndReturnMemObjPtr(maxTokenSize, hipDeviceMallocUncached);
   shmemStagingTokMemObj = ShmemMallocAndReturnMemObjPtr(maxTokenSize, hipDeviceMallocUncached);
